# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Grocery App",
    "description": "Grocery App - Flutter/NodeJS Series",
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "host": "localhost:4000",
  "basePath": "/",
...
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/api/slider": {
      "get": {
        "tags": ["Slider"],
        "summary": "Get all Slider",
        "parameters": [
          {
            "name": "sliderName",
            "in": "query",
            "required": false,
            "description": "Slider Name",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page Number",
            "type": "integer",
            "default": 1
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "description": "Page Size",
            "type": "integer",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SliderResponse"
            }
          }
        }
      },
      "post": {
        "tags": ["Slider"],
        "summary": "Create Slider API",
        "parameters": [
          {
            "name": "sliderName",
            "in": "formData",
            "description": "Slider Name",
            "required": true,
            "type": "string"
          },
          {
            "name": "sliderDescription",
            "in": "formData",
            "description": "Slider Description",
            "type": "string"
          },
          {
            "name": "sliderURL",
            "in": "formData",
            "description": "Slider URL",
            "type": "string"
          },
          {
            "name": "sliderImage",
            "in": "formData",
            "description": "Slider Image",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SliderResponse"
            }
          }
        }
      }
    },
    "/api/slider/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Slider Id",
          "type": "string"
        }
      ],
      "get": {
        "tags": ["Slider"],
        "summary": "Slider Details based on Slider Id",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SliderResponse"
            }
          }
        }
      },
      "put": {
        "tags": ["Slider"],
        "summary": "Update Slider API",
        "parameters": [
          {
            "name": "sliderName",
            "in": "formData",
            "description": "Slider Name",
            "required": true,
            "type": "string"
          },
          {
            "name": "sliderDescription",
            "in": "formData",
            "description": "Slider Description",
            "type": "string"
          },
          {
            "name": "sliderURL",
            "in": "formData",
            "description": "Slider URL",
            "type": "string"
          },
          {
            "name": "sliderImage",
            "in": "formData",
            "description": "Slider Image",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SliderResponse"
            }
          }
        }
      },
      "delete": {
        "tags": ["Slider"],
        "summary": "Delete Slider API",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SliderResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Slider": {
      "required": ["sliderName", "_id"],
      "properties": {
        "_id": {
          "type": "integer",
          "uniqueItems": true
        },
        "sliderName": {
          "type": "string"
        },
        "sliderDescription": {
          "type": "string"
        },
        "sliderURL": {
          "type": "string"
        },
        "sliderImage": {
          "type": "string"
        }
      }
    },
...
    "SliderResponse": {
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Slider"
          }
        }
      }
    }
  }
}